edomata_core::syntax

Trait ResultNecDecisionExt

pub trait ResultNecDecisionExt<T, R> {
    // Required methods
    fn to_decision_nec<E>(self) -> Decision<R, E, T>;
    fn to_accepted_nec(self) -> Decision<R, T, ()>;
}

Turns a Result with several errors into a decision.

Required Methods

fn to_decision_nec<E>(self) -> Decision<R, E, T>

Outputs the value or rejects with all the errors.

fn to_accepted_nec(self) -> Decision<R, T, ()>

Accepts the value as an event or rejects with all the errors.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".